+++ src/main/java/com/example/user/demo/service/UserService.java
@@ -4,17 +4,16 @@ import com.example.user.demo.binding.UserPayload;
 import com.example.user.demo.model.User;
 
 import java.util.List;
-import java.util.UUID;
 
 public interface UserService {
 
     public User create(UserPayload userPayload);
 
-    public User getOne(Integer id);
+    public User getOne(Long id);
 
     public List<User> getAll();
 
-    public User update(Integer id, UserPayload userPayload);
+    public User update(Long id, UserPayload userPayload);
 
-    public void delete(Integer id);
+    public void delete(Long id);
 }
